Added option to produce cumulative average.
Command-line option -a will cause Dprofiler to output a cumulative
average throughput rate (MB/s) when the -r (output rate) option is
also given.

The cumulative average is determined by adding the rate values (of
type double) to an accumulator after each group, then dividing by
the number of groups seen so far, in order to determine the average
rate to that point.

It is useful mainly for a "quick summary" of the results without
manual intervention or smoothing. Output is to a third column in
the resulting data file.